requirements.txt
pyserial
websocket-client
pandas
numpy
streamlit
scikit-learn==1.7.1
joblib
matplotlib